rag-server/pyproject.toml
[project]
name = "rag-server"
version = "1.0.0"
description = "RAG Pipeline Server for Voice AI"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "python-multipart>=0.0.20",
    "langchain>=0.3.0",
    "langchain-openai>=0.3.0",
    "langchain-text-splitters>=0.3.0",
    "langchain-qdrant>=0.2.0",
    "qdrant-client>=1.12.0",
    "openai>=1.60.0",
    "pypdf>=5.0.0",
    "python-docx>=1.1.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.7.0",
    "python-dotenv>=1.0.0",
    "structlog>=24.0.0",
    "httpx>=0.28.0",
    "redis[asyncio]>=5.0.0",
]
